警告の修正

+++ GameSources/Hunter.h
@@ -58,7 +58,7 @@ namespace basecross
 		@brief	村人のスピードの取得
 		*/
 		//--------------------------------------------------------------------------------------
-		int GetSpeed()
+		float GetSpeed()
 		{
 			return m_Speed;
 		}
@@ -68,7 +68,7 @@ namespace basecross
 		@brief	村人のスピードの設定
 		*/
 		//--------------------------------------------------------------------------------------
-		void SetSpeed(int VillagerSpeed)
+		void SetSpeed(float VillagerSpeed)
 		{
 			m_Speed = VillagerSpeed;
 		}
